+++ src/main/kotlin/com/provafacil/prova_facil/repository/SerieRepository.kt
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface SerieRepository : JpaRepository<Serie, Int> {
+interface SerieRepository : JpaRepository<Serie, Long> {
 }
